implementacion de cloudinary postmanga y correosfindall

+++ Servers/src/routes/Mangas/PostManga/index.ts
@@ -1,30 +1,46 @@
 import  { Router } from'express';
 import fs from 'fs-extra';
+import passport from "passport";
 import  Manga from '../../../models/Mangas/Manga.js';
+import User from "../../../models/Users/User.js";
 import { Uploadimage } from '../../../config/Cloudinary/index.js';
 import { FilesImage } from '../../../middlewares/FileUpload/index.js';
+import ReadTokenData from '../../../controles/Token/ReadTokenData/index.js';
 const router = Router();
 
-router.post('/', FilesImage(), async(req, res, next) => { 
-    const {title, genres, cover_image, description, rating, comments, chapter} = req.body;
-    try {  
-        if (req.files?.books){
-            const { books}: any = req.files;
-             let link: any = [];
-            let folderpath = `Mangas/${title}/chapter${chapter}`;
-            for (let i = 0; i < books.length; i++) {                
-                let linkClaudinary = await Uploadimage(books[i].tempFilePath, folderpath);
-                await fs.unlink(books[i].tempFilePath)
-                link.push(linkClaudinary.secure_url)
+router.post('/',  passport.authenticate("jwt", { session: false }), FilesImage(), async(req, res, next) => { 
+    const {title, genres, description, rating, comments, chapter} = req.body;
+    const {authorization} = req.headers;
+    try { 
+        const data= ReadTokenData(authorization);
+        const user = await User.findById(data.id);
+        if (user && user.admin){ 
+            if (req.files?.books){
+                let { books, cover_image }: any = req.files;
+                let link: any = [];
+                let folderpath = `Mangas/${title}/chapter${chapter}`;
+                for (let i = 0; i < books.length; i++) {                
+                    let linkClaudinary = await Uploadimage(books[i].tempFilePath, folderpath);
+                    await fs.unlink(books[i].tempFilePath)
+                    link.push(linkClaudinary.secure_url)
+                }
+                let mangas = {
+                    chapter: chapter,
+                    link: link
+                }
+                folderpath = `Mangas/${title}/cover_image`;
+                let CoverImgClaudinary = await Uploadimage(cover_image.tempFilePath, folderpath);
+                await fs.unlink(cover_image.tempFilePath);
+                cover_image = CoverImgClaudinary.secure_url;
+                const manga = new Manga({title, genres: JSON.parse(genres), cover_image, description, mangas, rating, comments});
+                let newmanga = await manga.save()
+                res.status(200).json(newmanga)
+            }else {
+                res.status(400).json("Informacion incompleta")
             }
-            let mangas = {
-               chapter: chapter,
-               link: link
-            }
-            const manga = new Manga({title, genres: JSON.parse(genres), cover_image, description, mangas, rating, comments});
-            let newmanga = await manga.save()
-            res.status(200).json(newmanga)
-        } 
+        } else{
+            res.status(400).json("No cuentas con autorizacion")
+        }
     } catch (error) {
         next(error)
     }
